WebScraperWithGPT-4o/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
openai==1.55.3
python-dotenv==1.0.1
streamlit==1.31.1
urllib3==2.0.7
fpdf2==2.7.8